[*] Reorganização das pastas e namespaces.
