+++ src/OpenAC.Net.Integrador/Commom/IntegradorEnvio.cs
@@ -1,12 +1,12 @@
-﻿// ***********************************************************************
-// Assembly         : OpenAC.Net.Integrador
-// Author           : marcosgerene
-// Created          : 07-17-2017
-//
-// Last Modified By : marcosgerene
-// Last Modified On : 07-17-2017
 // ***********************************************************************
-// <copyright file="IntegradorConstrutor.cs" company="OpenAC .Net">
+// Assembly         : OpenAC.Net.Integrador
+// Author           : RFTD
+// Created          : 05-30-2017
+//
+// Last Modified By : RFTD
+// Last Modified On : 05-30-2017
+// ***********************************************************************
+// <copyright file="IntegradorEnvio.cs" company="OpenAC .Net">
 //		        		   The MIT License (MIT)
 //	     		    Copyright (c) 2014 - 2021 Projeto OpenAC .Net
 //
@@ -29,27 +29,31 @@
 // <summary></summary>
 // ***********************************************************************
 
-using OpenAC.Net.Core.Generics;
 using OpenAC.Net.DFe.Core.Attributes;
+using OpenAC.Net.DFe.Core.Document;
 
-namespace OpenAC.Net.Integrador
+namespace OpenAC.Net.Integrador.Commom
 {
-    public sealed class IntegradorConstrutor : GenericClone<IntegradorConstrutor>
+    [DFeRoot("Integrador")]
+    public sealed class IntegradorEnvio : DFeDocument<IntegradorEnvio>
     {
         #region Constructors
 
-        public IntegradorConstrutor()
+        public IntegradorEnvio()
         {
-            Parametros = new IntegradorParametroCollection();
+            Identificador = new IntegradorIdentificador();
+            Componente = new IntegradorComponente();
         }
 
         #endregion Constructors
 
         #region Properties
 
-        [DFeCollection("Parametros")]
-        [DFeItem(typeof(IntegradorParametro), "Parametro")]
-        public IntegradorParametroCollection Parametros { get; set; }
+        [DFeElement("Identificador", Ocorrencia = Ocorrencia.Obrigatoria)]
+        public IntegradorIdentificador Identificador { get; set; }
+
+        [DFeElement("Componente", Ocorrencia = Ocorrencia.Obrigatoria)]
+        public IntegradorComponente Componente { get; set; }
 
         #endregion Properties
     }
